refactor(types): split journal.ts into entities/{journal-entry,wiki-page}.ts
