+++ packages/types/src/__tests__/encryption.test.ts
@@ -70,6 +70,7 @@ import type { Channel } from "../entities/channel.js";
 import type { CustomFront } from "../entities/custom-front.js";
 import type { FrontingComment } from "../entities/fronting-comment.js";
 import type { FrontingSession } from "../entities/fronting-session.js";
+import type { JournalEntry } from "../entities/journal-entry.js";
 import type { LifecycleEvent } from "../entities/lifecycle-event.js";
 import type { MemberPhoto } from "../entities/member-photo.js";
 import type { Member } from "../entities/member.js";
@@ -79,6 +80,7 @@ import type { PollVote } from "../entities/poll-vote.js";
 import type { Poll } from "../entities/poll.js";
 import type { Relationship } from "../entities/relationship.js";
 import type { TimerConfig } from "../entities/timer-config.js";
+import type { WikiPage } from "../entities/wiki-page.js";
 import type { Group } from "../groups.js";
 import type {
   BucketId,
@@ -91,7 +93,6 @@ import type {
   SlugHash,
   SystemId,
 } from "../ids.js";
-import type { JournalEntry, WikiPage } from "../journal.js";
 import type { UnixMillis } from "../timestamps.js";
 import type { Serialize } from "../type-assertions.js";
